DiscoFish Player is now TechnoGecko Light
A lot of things has changed. ProjectM patterns are gone (we were not using
them for the last couple of years anyway), all patterns are now generated by
the same code that we were using on wearables. Kinect code is gone because
we no longer have kinect on the vehicle (but if we put it back, the code
is here in commit history, and also on a separate branch).

Most of the pattern generaton code was moved to a separate library
(https://github.com/unisparks/unisparks). What's left here is a thin
configuration layer around Unisparks core plus web/phone control app.